Added a function to reset all the brains starting activations
You can see that this was necessary because if you took a single brain and tested it multiple times (without mutations) on the same maze, it scored very differently each time to my surprise.
The reason is that once it's finished the maze once, it's starting with a different "state of mind" the next time.  It could end up solving the maze 4x as fast the second time around
despite being the exact same brain.  That just goes to show how much probability has to do with this task.  Not good...